+++ packages/zowe-explorer-api/src/profiles/UserSettings.ts
@@ -11,13 +11,14 @@
 
 import { ICreateDataSetOptions } from "@zowe/cli";
 
-export interface CreateDSOptionsWithTemplate extends ICreateDataSetOptions {
-    /**
-     * Name of data creation template
-     */
-    name?: string;
-}
-export type DataSetAllocTemplate = CreateDSOptionsWithTemplate;
+/**
+ * @type DataSetAllocTemplate
+ * Used during creation of data sets in Zowe Explorer to access and manipulate a list of saved templates
+ * presented to or saved by the user
+ */
+export type DataSetAllocTemplate = {
+    [key: string]: ICreateDataSetOptions;
+};
 
 /**
  * The types of persistence schemas wich are available in settings.json